Sizes are unsigned

+++ test/default/generichash2.c
@@ -10,7 +10,7 @@ int main(void)
 #define MAXLEN 64
     crypto_generichash_state st;
     crypto_uint8 in[MAXLEN], out[crypto_generichash_BYTES_MAX], k[crypto_generichash_KEYBYTES_MAX];
-    int h,i,j;
+    size_t h,i,j;
 
     for(h = 0; h < crypto_generichash_KEYBYTES_MAX; ++h) k[h] = h;
